Add setUserData mutation and consolidate pledge tests
Add GraphQL mutation to set key-value pairs in PledgeUser's user_data
field. Also merge mutation tests from test_pledge_schema.py into
test_graphql_pledge.py to keep all pledge GraphQL tests in one file.
